requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
boto==2.48.0
boto3==1.7.4
botocore==1.10.4
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
DateTime==4.2
Django==2.0.4
django-cors-headers==2.2.0
docopt==0.6.2
docutils==0.14
gast==0.2.0
gensim==3.4.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
jmespath==0.9.3
Keras==2.1.6
Markdown==2.6.11
mwapi==0.5.1
mwcli==0.0.2
mwviews==0.1.0
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
para==0.0.6
pipreqs==0.4.9
protobuf==3.5.2.post1
python-dateutil==2.6.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
scipy==1.0.1
six==1.11.0
smart-open==1.5.7
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
yarg==0.1.9
zope.interface==4.4.3
google-cloud-bigquery==0.31.0
google-auth-oauthlib==0.2.0